Switch integer types in serial console query parameters to u64, better match for underlying usize

+++ common/src/api/internal/nexus.rs
@@ -110,15 +110,15 @@ pub struct InstanceSerialConsoleRequest {
     /// Character index in the serial buffer from which to read, counting the bytes output since
     /// instance start. If this is not provided, `most_recent` must be provided, and if this *is*
     /// provided, `most_recent` must *not* be provided.
-    pub from_start: Option<u32>,
+    pub from_start: Option<u64>,
     /// Character index in the serial buffer from which to read, counting *backward* from the most
     /// recently buffered data retrieved from the instance. (See note on `from_start` about mutual
     /// exclusivity)
-    pub most_recent: Option<u32>,
+    pub most_recent: Option<u64>,
     /// Maximum number of bytes of buffered serial console contents to return. If the requested
     /// range runs to the end of the available buffer, the data returned will be shorter than
     /// `max_bytes`.
-    pub max_bytes: Option<u32>,
+    pub max_bytes: Option<u64>,
 }
 
 /// Contents of an Instance's serial console buffer.
@@ -129,5 +129,5 @@ pub struct InstanceSerialConsoleData {
     pub data: Vec<u8>,
     /// The absolute offset since boot (suitable for use as `byte_offset` in a subsequent request)
     /// of the last byte returned in `data`.
-    pub last_byte_offset: u32,
+    pub last_byte_offset: u64,
 }
